requirements.txt
# Core AI / LangChain
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.1.4
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
openai==1.109.1

# Vector DB & Embeddings
chromadb==1.5.5
sentence-transformers==5.3.0
transformers==4.46.3
torch==2.11.0

# Alibaba Cloud PAI DSW SDK
alibabacloud-pai-dsw20220101==2.4.0
alibabacloud-tea-openapi==0.4.4
alibabacloud-tea==0.4.3
alibabacloud-tea-util==0.3.14
alibabacloud-credentials==1.0.8
alibabacloud-credentials-api==1.0.0
alibabacloud_gateway_spi==0.0.3
darabonba-core==1.0.5

# Storage & Infrastructure
redis==7.4.0
chromadb==1.5.5

# Web
Flask==3.1.3

# System & Data
psutil==7.2.2
pandas==3.0.1

# Env
python-dotenv
